Separating Platforms
Hey guys, I mentioned this to Allison but I'm having some trouble separating a few of the platforms. I did a char search and for some of the categories, the script groups together any platform that is similar in platform name (like instead of separating 'PlayStation', it will also include 'PlayStation 2, 3, 4, Vita' etc. If you have an idea of how to fix it that would be awesome!